# server/src/main/resources/open-api-contract.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenFireStationManager API",
...
    "contact": {
      "name": "OpenFireStationManager",
      "url": "https://github.com/simonhauck/OpenFireStationManager"
    },
    "version": "v1"
  },
  "security": [ {
    "sessionCookie": [ ]
  } ],
...
  "paths": {
...
    "/api/clothing/locations": {
      "get": {
        "tags": [ "Clothing Locations" ],
        "summary": "List every clothing location",
        "description": "Returns all places garments can be kept (`Standort`), including personal lockers, the shared pool, and the laundry. Start here to discover valid `locationId` values for checkout, relocation, and reconciliation.\n\nThis endpoint returns locations flagged `onlyVisibleForKleiderwart` to every caller; it is the item lookups, not this list, that apply that visibility rule.\n\n**Authorisation:** any signed-in user.",
        "operationId": "listClothingLocations",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClothingLocation"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated — no valid session cookie was supplied.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error. The response body carries no details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [ "Clothing Locations" ],
        "summary": "Create a new clothing location",
        "description": "Registers a new place where garments can be kept. Choose the `type` carefully: it determines whether the location behaves as shared stock (`POOL`), as a laundry staging area (`WAESCHE`), or as one member's personal storage (`PERSONAL`).\n\n**Authorisation:** requires the `KLEIDERWART` role. An `ADMIN` account implicitly holds every role and may also call this.",
        "operationId": "createClothingLocation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClothingLocationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The newly created location.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ClothingLocation"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated — no valid session cookie was supplied.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but the account lacks the required role.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error. The response body carries no details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "400": {
            "description": "The request failed validation. The `errors` array names each offending field.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/clothing/locations/batch": {
      "post": {
        "tags": [ "Clothing Locations" ],
        "summary": "Create many clothing locations in one request",
        "description": "Creates several locations at once — the usual way to set up a whole row of lockers in a single call.\n\n**Authorisation:** requires the `KLEIDERWART` role. An `ADMIN` account implicitly holds every role and may also call this.",
        "operationId": "createClothingLocationsBatch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchCreateClothingLocationsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The newly created locations.",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClothingLocation"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated — no valid session cookie was supplied.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but the account lacks the required role.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error. The response body carries no details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "400": {
            "description": "The request failed validation. The `errors` array names each offending field.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetail"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/clothing/locations/{id}": {
      "get": {
        "tags": [ "Clothing Locations" ],
        "summary": "Get one clothing location by its numeric id",
        "description": "Looks up a single location by its database id. This returns the location's own properties only — it does not list the garments currently stored there. To find those, use `searchClothingItems`.\n\n**Authorisation:** any signed-in user.",
        "operationId": "getClothingLocation",
        "parameters": [ {
          "name": "id",
          "in": "path",
          "description": "Numeric id of the clothing location.",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64",
            "exclusiveMinimum": 0
          },
          "example": 7
        } ],
        "responses": {
          "200": {
            "description": "The clothing location.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ClothingLocation"
                }
              }
            }
          },
          "404": {
            "description": "No clothing location exists with this id.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated — no valid session cookie was supplied.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error. The response body carries no details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "400": {
            "description": "The request failed validation. The `errors` array names each offending field.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetail"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [ "Clothing Locations" ],
        "summary": "Delete a clothing location",
        "description": "Removes a place from the station. Move any garments still stored here first — deleting a location does not delete or relocate its contents, and items left pointing at it become hard to find.\n\n**Authorisation:** requires the `KLEIDERWART` role. An `ADMIN` account implicitly holds every role and may also call this.",
        "operationId": "deleteClothingLocation",
        "parameters": [ {
          "name": "id",
          "in": "path",
          "description": "Numeric id of the clothing location to delete.",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64",
            "exclusiveMinimum": 0
          },
          "example": 7
        } ],
        "responses": {
          "204": {
            "description": "The location was deleted."
          },
          "404": {
            "description": "No clothing location exists with this id.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated — no valid session cookie was supplied.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but the account lacks the required role.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error. The response body carries no details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "400": {
            "description": "The request failed validation. The `errors` array names each offending field.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetail"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [ "Clothing Locations" ],
        "summary": "Replace the details of a clothing location",
        "description": "Despite the `PATCH` verb this is a **full replacement**: `name`, `comment`, `type`, and `onlyVisibleForKleiderwart` are all applied from the request body. Read the location first and resend the fields you want to keep.\n\nGarments stored here are unaffected and stay where they are.\n\n**Authorisation:** requires the `KLEIDERWART` role. An `ADMIN` account implicitly holds every role and may also call this.",
        "operationId": "updateClothingLocation",
        "parameters": [ {
          "name": "id",
          "in": "path",
          "description": "Numeric id of the clothing location to replace.",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64",
            "exclusiveMinimum": 0
          },
          "example": 7
        } ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClothingLocationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The updated location.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ClothingLocation"
                }
              }
            }
          },
          "404": {
            "description": "No clothing location exists with this id.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated — no valid session cookie was supplied.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but the account lacks the required role.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error. The response body carries no details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "400": {
            "description": "The request failed validation. The `errors` array names each offending field.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetail"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/clothing/items/search": {
      "get": {
        "tags": [ "Clothing Items" ],
        "summary": "Search clothing items by type name, size, or barcode",
        "description": "Free-text search across the clothing type name, the size, and the barcode. Use this to answer questions like \"which jackets in size 52 do we have?\" — pass the type name and the size together as one query string.\n\nResults are returned in **resolved** form, each embedding its full clothing type and current location. Callers without the `KLEIDERWART` role never see items held in locations flagged `onlyVisibleForKleiderwart`.\n\nThe result set is capped at 50 entries regardless of the `limit` supplied, so a full result set is never guaranteed; narrow the query rather than raising the limit.\n\n**Authorisation:** any signed-in user.",
        "operationId": "searchClothingItems",
        "parameters": [ {
          "name": "q",
          "in": "query",
          "description": "Search term matched against clothing type name, size, and barcode. Required, and must not be blank.",
          "required": true,
          "schema": {
            "type": "string"
          },
          "example": "Einsatzjacke 52"
        }, {
          "name": "limit",
          "in": "query",
          "description": "Maximum number of results to return. Values above the server-side cap of 50 are silently reduced to 50.",
          "required": false,
          "schema": {
            "type": "integer",
            "format": "int32",
            "default": 50
          },
          "example": 50
        } ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResolvedClothingItem"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated — no valid session cookie was supplied.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error. The response body carries no details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "400": {
            "description": "The request failed validation. The `errors` array names each offending field.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetail"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ProblemDetail": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string",
            "format": "uri"
          },
          "properties": {
            "type": "object",
            "additionalProperties": { }
          }
        }
      },
      "ValidationError": {
        "description": "A single field-level validation failure.",
        "properties": {
          "field": {
            "type": "string",
            "description": "Path of the offending field within the request body, e.g. `size`. Empty when the rule applies to the request as a whole rather than one field.",
            "example": "size"
          },
          "message": {
            "type": "string",
            "description": "Human-readable reason the value was rejected.",
            "example": "must not be blank"
          },
          "rejectedValue": {
            "type": "object",
            "description": "The value that was submitted and rejected. May be null."
          }
        }
      },
      "ValidationProblemDetail": {
        "description": "RFC 9457 problem document describing a request that failed validation, extended with an `errors` array naming each offending field.",
        "properties": {
          "type": {
            "type": "string",
            "example": "about:blank"
          },
          "title": {
            "type": "string",
            "example": "Invalid request"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "example": 400
          },
          "detail": {
            "type": "string",
            "example": "Validation failed for request body"
          },
          "instance": {
            "type": "string",
            "description": "URI of the request that failed, when available."
          },
          "errors": {
            "type": "array",
            "description": "Every field-level failure detected in this request.",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
...
      "EntityMetaData": {
        "type": "object",
        "description": "Audit trail attached to every stored record. Maintained by the server on save — these fields are ignored if supplied in a request. `System` appears as the actor for changes made outside a user session, such as database migrations.",
        "properties": {
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the record was first stored."
          },
          "createdBy": {
            "type": "string",
            "description": "Username of whoever created the record, or `System`.",
            "example": "m.mustermann"
          },
          "lastModifiedAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the record was last changed."
          },
          "lastModifiedBy": {
            "type": "string",
            "description": "Username of whoever last changed the record, or `System`.",
            "example": "m.mustermann"
          }
        },
        "required": [ "createdAt", "createdBy", "lastModifiedAt", "lastModifiedBy" ]
      },
...
      "ClothingItem": {
        "type": "object",
        "description": "One physical garment. `typeId` and `locationId` are numeric references, not embedded objects; see `ResolvedClothingItem` for the form that includes them in full.",
        "properties": {
          "typeId": {
            "type": "integer",
            "format": "int64",
            "description": "Id of the clothing type this garment belongs to.",
            "example": 3
          },
          "size": {
            "type": "string",
            "description": "Size label as printed on the garment.",
            "example": "52"
          },
          "barcode": {
            "type": "string",
            "description": "Barcode printed on the garment, unique across all items. Null if untagged.",
            "example": "1234567890128"
          },
          "locationId": {
            "type": "integer",
            "format": "int64",
            "description": "Id of the location holding this garment. Null when unassigned.",
            "example": 7
          },
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Server-assigned identifier.",
            "example": 42
          },
          "metaData": {
            "$ref": "#/components/schemas/EntityMetaData"
          }
        },
        "required": [ "id", "metaData", "size", "typeId" ]
      },
      "CreateClothingLocationRequest": {
        "type": "object",
        "description": "Full set of properties for a clothing location. Used for both creation and update; on update every field is applied, so this is a full replacement.",
        "properties": {
          "name": {
            "type": "string",
            "description": "Display name of the place, e.g. `Spind 12` or `Lager Regal B`.",
            "example": "Spind 12",
            "maxLength": 255,
            "minLength": 0
          },
          "comment": {
            "type": "string",
            "description": "Free-text note about this location. Required, but may be an empty string — it is not nullable.",
            "example": "Zweite Reihe, links",
            "maxLength": 255,
            "minLength": 0
          },
          "onlyVisibleForKleiderwart": {
            "type": "boolean",
            "description": "When true, this location and every garment in it are hidden from users who do not hold the `KLEIDERWART` role. Such items read as not found in barcode lookup and never appear in search results.",
            "example": false
          },
          "type": {
            "type": "string",
            "description": "What kind of place this is.",
            "enum": [ "POOL", "WAESCHE", "PERSONAL", "OTHER" ]
          }
        },
        "required": [ "comment", "name", "onlyVisibleForKleiderwart", "type" ]
      },
      "ClothingLocation": {
        "type": "object",
        "description": "A place where garments are kept (`Standort`) — the shared pool, the laundry, a member's personal locker, or anywhere else.",
        "properties": {
          "name": {
            "type": "string",
            "description": "Display name of the place.",
            "example": "Spind 12"
          },
          "comment": {
            "type": "string",
            "description": "Free-text note. May be empty, never null."
          },
          "onlyVisibleForKleiderwart": {
            "type": "boolean",
            "description": "When true, only `KLEIDERWART` users may see this location and its contents.",
            "example": false
          },
          "type": {
            "type": "string",
            "description": "What kind of place this is.",
            "enum": [ "POOL", "WAESCHE", "PERSONAL", "OTHER" ]
          },
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Server-assigned identifier.",
            "example": 7
          },
          "metaData": {
            "$ref": "#/components/schemas/EntityMetaData"
          }
        },
        "required": [ "comment", "id", "metaData", "name", "onlyVisibleForKleiderwart", "type" ]
      },
      "BatchCreateClothingLocationsRequest": {
        "type": "object",
        "description": "A set of clothing locations to create together in a single request.",
        "properties": {
          "items": {
            "type": "array",
            "description": "Locations to create. Must contain at least one entry, and every entry is validated individually.",
            "items": {
              "$ref": "#/components/schemas/CreateClothingLocationRequest"
            },
            "minItems": 1
          }
        },
        "required": [ "items" ]
      },
...
      "ResolvedClothingItem": {
        "type": "object",
        "description": "A clothing item with its type and current location embedded in full, so no follow-up lookups are needed to display it.",
        "properties": {
          "clothingItem": {
            "$ref": "#/components/schemas/ClothingItem",
            "description": "The garment itself."
          },
          "location": {
            "$ref": "#/components/schemas/ClothingLocation",
            "description": "Where the garment currently is. Null when the item is not assigned to any location."
          },
          "clothingType": {
            "$ref": "#/components/schemas/ClothingType",
            "description": "The type (`Kleidungsart`) this garment belongs to."
          }
        },
        "required": [ "clothingItem", "clothingType" ]
      },
...
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "description": "Server-side session cookie issued by `POST /api/public/auth/login`. Send it on every authenticated request.",
        "name": "OFSM_AUTH",
        "in": "cookie"
      },
      "rememberMeCookie": {
        "type": "apiKey",
        "description": "Long-lived cookie issued when logging in with `rememberMe = true`, valid for 30 days. It authenticates on its own, so a long-lived client may send it instead of a session cookie.",
        "name": "OFSM_AUTH_REMEMBER_ME",
        "in": "cookie"
      }
    }
  }
}